feat: Acquire a tokio-postgres connection on startup
Add tokio-postgres as a dependency and acquire a connection on startup.
For now, we assume that postgres is created with the default user
(`postgres`), the password `password`, is running locally
(`host=localhost`), without TLS, and on the default port (`5432`).

Initialize a tokio-postgres connection when starting the server, and
spawn a `conn.await` task. Spawning this task is necessary for
communication with the database to actually occur. The future will only
resolve when the connection is closed, either because an error occurred
or because the associated client dropped and all outstanding work
completed. This means that the existing task will simply print the error
if one occurs, but it will not recover from it.

Add documentation to the `README.md` on how to spin up a PostgreSQL
container with the necessary configuration for pasta6 to connect to it.