feat: Create `pastes` table in DB on startup
Add an `init_db` method which creates the `pastes` table in the DB if
the table is not already present. `init_db` is run on startup, before
any routes are initialized. This should ensure that if the warp routes
are being served, the database has been correctly initialized. The DB
initialization script creates 3 columns: `id`, `name`, and `created_at`.
`id` is an incrementing primary key, `name` represents the identifier
that the paste can be accessed by, and `created_at` is a timestamp of
when the paste was created.