impr: Change default DB username, password, and name
Change the default DB username, password, and database name to
`pastaaaaaa` for everything. Update documentation to show how to
initialize a PostgreSQL database with those settings so pasta6 can
connect to it. Using non-default values for these parameters will make
it easier to change the documentation later to explain how to set this
up on a production EC2 instance.