feat: Add DB connection check to healthcheck
Add a `SELECT 1` query to the PostgreSQL database for when the `/health`
endpoint is accessed. This allows us to periodically query the `/health`
endpoint and restart the service if we experienced a DB connection
interruption. A supervisor needs to be added to continuously monitor this
endpoint.

Add a `Error` enum for wrapping possible errors. Currently the enum
provides a `DbQueryError` variant which wraps `tokio_postgres::Error`.
Wrapping this error allows our healthcheck endpoint to debug print the
error, if one occurs. Ideally we should provide a proper JSON
representation of the error.