feat: Add a REST `POST /paste` endpoint
Add a REST `POST /paste` endpoint which takes arbitrary bytes in the
HTTP request body (up to a content length of 16kb) and stores them in
the database. Also stored in the database is the timestamp of when the
paste was created (i.e. when the record was inserted into the database)
as well as an auto-incrementing ID representing the primary key of the
record.

The response to the paste request is a JSON message containing an `"id"`
field which is the ID of the paste in the database. Here is an example
request/response pair:

    $ curl -XPOST -dhello localhost:3030/paste
    {"id":1}

In the future, the primary key of pastes should not be returned to the
user as this leaks internal information, as attackers now know that
paste keys are sequential and thus they know how many pastes are
currently in the database.

Add `chrono` library as a dependency to provide the `DateTime<Utc>` type
for (de)serialization of timestamps of REST JSON paste create requests and paste
data in the database.