feat: Add a REST `GET /paste/<id>` endpoint
Add a REST `GET /paste/<id>` endpoint which takes an integer paste ID as
a URL param and returns the corresponding paste after fetching it from
the DB. Currently, if no paste is found by that ID then a 400 "Could not
execute request" is returned, which should probably be a 404 instead.
Passing in a non-integer paste ID results in a "Method not allowed"
which could probably be improved by returning a 400 instead.

To lookup a paste, one could do the following:

    $ curl localhost:3030/paste/1
    hello