impr: Add deadpool_postgres connection pool
Add `deadpool_postgres` library to provide a connection pool for
`tokio_postgres` DB connections. We want to use a connection pool to
avoid contention on a single connection to the DB.

Add `PG_HOST`, `PG_USER`, and `PG_PASSWORD` environment variables which
configure our connection to the postgres database. These environment
variables must be set on startup or the service will exit.