impr: Add nginx support for serving static files
Add `default.conf` which configures nginx to proxy all `/api/`
requests to pasta6. All other requests will be served as static files
directly by nginx. This is done so that we can serve files like
`index.html` and CSS without making a request to pasta6.

Make the server host configurable via `PASTA6_HOST` env var. The
`PASTA6_HOST` env var must be set to an IPv4 address on start-up. pasta6
will bind to the provided address. This address can be `127.0.0.1` or
`localhost` to for security purposes, but proxies like Nginx must then
be on the same machine to forward requests correctly.

If, for example, you are running pasta6 on the host directly (e.g. for
development) but nginx in a container, you'll need to set `PASTA6_HOST`
to the IPv4 address of the host machine, and you'll need to modify the
upstream IP address in `default.conf` to point at the same IP. This way,
nginx will be able to proxy requests to pasta6 successfully.