impr: Add HTML templating via askama, and CSS via TailwindCSS
Add `askama` and `askama_warp` as dependencies for providing server-side
run-time templating of webpages. Generating responses from templates on
the server allows us to dynamically display content (e.g. pastes) without
requiring JavaScript (i.e. to make XHR requests and then display them).

We add both the `askama` and `askama_warp` dependencies due to an
outstanding issue where importing `askama::Template` does not work.
Importing `askama_warp::Template` (which is the same object) is the
suggested workaround for now.
See: https://github.com/djc/askama/issues/346

Add `tailwindcss` as a yarn dependency. TailwindCSS is used as the CSS
framework for the website. It was chosen so that I could get some
familiarity with a modern CSS framework, and because it is oriented
around allowing you to design the page yourself, instead of providing
pre-packaged components. Pre-packaged components would save on
development time but I want to try my hand at website design and
server-side component re-use (which is why I added `askama` as a
dependency).

`yarn` is used to compile the CSS to deployable static files.
Currently, compiling the CSS is a manual step:

    yarn run tailwindcss build styles.css -o static/styles.css

`NODE_ENV=production` can be set when running the above command so that
TailwindCSS removes unused CSS elements. This greatly reduces the size
of the resulting CSS file (~2.4MB -> ~17KB).

nginx must be configured to serve the CSS as a sstatic file (instead of
the warp server serving it), so the `README.md` has been updated with an
example of the new production nginx config. An `upstream` section has
been added which points at the local pasta6 instance. Most requests are
still proxied to pasta6, but `styles.css` is served as a static file
from the `/srv/www/p.dank.xyz` directory.

Add a `PASTA6_HOST` environment variable for configuring the IPv4
address pasta6 binds to. This value should be set to your local
machine's IP in development (e.g. `PASTA6_HOST=172.16.1.25`). This is
necessary so that nginx can proxy requests to pasta6. If you set the
host IP to `127.0.0.1` and nginx is running in a container then pasta6
will reject requests as they are not arriving on the correct IP.

In production, `PASTA6_HOST` should be set to `127.0.0.1` to ensure that
external requests are not allowed to reach pasta6, and that they _must_
be proxied through nginx.

Re-write the `with_db` filter to fetch a client connection for the user
instead of providing the pool. There is currently no reason for handlers
to talk to the pool (as they just need it to get a connection), and
doing it this way deduplicates some error handling.

Move the existing paste endpoints behind `/api/paste` and add new
`/paste` endpoints which are used to display an HTTP UI for the paste
service. These endpoints render the new paste page using the
`templates/index.html` template and the get paste page using the
`templates/paste.html` template. This is similar to traditional Flask
apps in that we can provide functionality through GET/POST and HTTP
forms. Currently we use `application/x-www-form-urlencoded` forms
instead of `multipart/form-data` as the latter is intended for file
uploads whereas we are currently dealing with strings.

Refactor the routes to a `routes` module. This takes up a few more lines
of code but lets us organize the route definitions separately from
server initialization, which will be helpful later on when we run
several applications off of a single instance.

Replace `static/index.html` with a `templates/index.html` askama
template.

Configure TailwindCSS with a `tailwind.config.js` file which currently
just enabled backwards-incompatible changes
(`removeDeprecatedGapUtilities: true`) and tells tailwind where to look
for HTML files so that we can slim down the compiled CSS -- via the
`purge` and `plugins` configuration options.