feat: Add auth module
Add an empty auth module. This module will provide
registration and authentication functionality, for ensuring the privacy
of pastes.

Currently, the auth module just contains a `GET /register` endpoint
which serves a content-less `register.html` template.

Add `Register` button which links to the `/register` page to the global
nav bar.