feat: Add Makefile
Replace `PG_` prefix for environment variables with `POSTGRES_` to make
it consistent with the environment variables used for the postgres
container itself.

Add a `POSTGRES_DB` environment variable for setting the postgres DB
name.

Move the installation/quickstart commands from the `README.md` to a
`Makefile` so that they can be execute more easily.