feat: Add registration endpoint
Add `rand` and `serde_json` as dependencies. `rand` is used to generate
random session IDs (but session IDs are not encrypted or securely random
yet). `serde_json` is used to deserialize/serialize the session ID from/into
the `session` cookie (though perhaps this should be base64-encoded).

Change postgres docker container command to `postgres -c
log_statement=all` instead of `postgres` to ensure that all statements
are logged. This is useful for debugging queries.

Add a `session` module which provides warp filters for deserializing the
`session` cookie into a `Session` struct which is a newtype wrapping a
`String`. The `String` wrapped in the `Session` struct represents a
"session ID" which is currently a randomly-generated alphanumeric
string. The session ID is a unique value that is stored in the DB and
provides stateful authentication. When a user performs a request, the
session is read from the cookie and matched against the sessions in the
DB; if there is a match the user is treated as logged in.
If we fail to deserialize a cookie, or fetch it from the DB, etc. then
all these errors are hidden from the user and are treated as if the user
was not logged in.

Add a `Logged in as {{ username }}` text span to the navbar that appears
when the user is logged in.
Hide the `Register` button in the navbar if the user is currently logged
in, and replace it with a `Logout` button that currentl doesn't do
anything.

Create a `TemplateContext` type which is used to store variables shared
between all templates (i.e. any template that inherits from
`base.html`). Currently this is just a `current_user` field, which is an
optional `User`, provided if the user is currently logged in.

Expand the `auth` module to provide basic registration functionality.
Add a `POST /register` endpoint which creates a user and a session in
the DB on a successful registration request.