refr: Refactor user postgres operations to a `PostgresStore`
Add `async_trait` crate to provide `async` fn support in traits.

Create a `User` trait to define the interface of the user object, along
with a `UserStore` trait that defines the operations on a backing store
that are necessary to provide the `User` API. Refactor the existing DB
calls for user operations to the new `PostgresStore` implementation of
`UserStore`.

Limit visibility of all `pub` items to `pub(crate)`.