tmp: add Sentry
Add initial Sentry integration. Currently we use the default
configuration, except for setting `debug` to true if we've compiled our
project without `debug_assertions` enabled. This should only be the case
for `--release` builds.

Add a `pasta6.sentry_dsn` global config key for getting the sentry DSN
value to use for the application.

Configure Sentry to log the release version as
"$CARGO_BIN_NAME@$(git describe --always --abbrev=40 --dirty)".
This ensures that we can tell which binary emitted the error/panic
as well as what the latest commit was. The hash will be suffixed with
`-dirty` if the workspace is not clean. This is intentional so that we
can distinguish WIP builds from proper releases.