impr: Catch panics in worker threads & send 500 response
Uncaught panics in a request handler result in a terminated request.
In our case, this means that nginx returns an upstream proxy error. We'd
like to have some more control over this scenario, so we wrap the warp
service that we create with `warp::service(routes)` in a `handle_panics`
function.

`handle_panics` is a function that takes a future (our request handler)
and catches any panics it throws. It then logs an error (but not any
information about the panic) and returns a 500 response to the user. The
panic still gets logged (by Tokio?) so we don't lose the backtrace in
development environments and gets picked up by Sentry regardless. This
just provides a slightly improved experience for the user.

Ideally, we'd display a templated HTML page to the user in this case.

There are a few caveats to this method. In particular, we are using
`catch_unwind` from `futures_util` instead of `std`. This is because
futures are lazy and we want to catch the unwinds from its `poll`
method, not from the future being created.
The future is likely not going to be unwind-safe.
For more info, see: https://vorner.github.io/2020/04/13/hyper-traps.html